25.01 devel clean internal git 2 (#1255)

+++ .github/container/git-clone.sh
@@ -79,7 +79,9 @@ COMMIT_SHA=$(git rev-parse HEAD)
 git submodule update --init --recursive
 if [[ "${GIT_REPO}" == *"gitlab"* ]]; then
   git remote remove origin
-  grep -r gitlab-ci-token .git | awk -F: '{print $1}' | xargs rm -f
+  if grep -q -r gitlab-ci-token .git; then
+    grep -r gitlab-ci-token .git | awk -F: '{print $1}' | xargs rm -f
+  fi
 fi
 popd
